requirement.txt
absl-py==2.3.1
alembic==1.16.5
anyio==4.12.1
appdirs==1.4.4
astunparse==1.6.3
backoff==2.2.1
blinker==1.9.0
boto3==1.42.39
botocore==1.42.39
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
commonmark==0.9.1
contourpy==1.3.0
cryptography==46.0.4
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.19
dataclasses-json==0.6.7
dnspython==2.7.0
docker==7.1.0
entrypoints==0.4
exceptiongroup==1.3.1
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.60.2
fusepy==3.0.1
gast==0.7.0
gensim==4.3.3
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-pasta==0.2.0
gql==4.0.0
graphql-core==3.2.7
greenlet==3.2.4
grpcio==1.76.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.11
imbalanced-learn==0.12.4
importlib_metadata==7.2.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
Mako==1.3.10
Markdown==3.9
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.9.4
ml-dtypes==0.2.0
mlflow==2.10.2
-e 
multidict==6.7.1
mypy_extensions==1.1.0
nltk==3.8.1
numpy==1.24.3
oauthlib==3.3.1
opt_einsum==3.4.0
optuna==4.6.0
packaging==23.2
pandas==2.1.4
pathvalidate==3.0.0
pillow==11.3.0
prometheus-client==0.19.0
propcache==0.4.1
protobuf==4.25.8
pyarrow==15.0.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pymongo==4.6.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2023.4
pywin32==311
PyYAML==6.0.3
querystring-parser==1.2.4
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.1.0
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.3.2
scipy==1.13.1
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlparse==0.5.5
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.4.0
tqdm==4.67.2
treelib==1.6.4
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
waitress==2.1.2
Werkzeug==3.1.5
wrapt==1.14.2
xgboost==1.7.6
yarl==1.22.0
zipp==3.23.0